requirements.txt
langchain
langchain-community
langgraph
langchain-text-splitters
chromadb
sentence-transformers
pypdf
python-dotenv
requests